feat: Add Feishu and DingTalk document loaders with bug fixes
New Features:
- Add Feishu document loader with single document and space batch loading support
- Add DingTalk document loader with single document and workspace batch loading support
- Implement comprehensive test coverage for both loaders
- Support API authentication and token management
- Handle pagination for batch document loading

Bug Fixes:
- Fix 3 compilation errors in core module (TempMemory, TopicTrackingMemory)
- Fix YuqueService baseUrl NPE issue (uninitialized static field)
- Fix missing modelVersion in wecom module pom.xml

Code Quality:
- Add Apache 2.0 License headers to Yuque related files
- Remove unused imports
- Set author to Libres-coder for all new files
- Follow existing project code style and architecture

Testing:
- All tests passing (6 tests total: 3 for Feishu, 3 for DingTalk)
- No breaking changes
- Existing functionality remains intact

+++ ali-langengine/ali-langengine-core/src/main/java/com/alibaba/langengine/core/memory/impl/TempMemory.java
@@ -533,7 +533,7 @@ public class TempMemory extends BaseMemory {
         Map<String, Long> categoryDistribution = categoryIndex.entrySet().stream()
             .collect(Collectors.toMap(
                 Map.Entry::getKey,
-                entry -> entry.getValue().size()
+                entry -> (long) entry.getValue().size()
             ));
         stats.put("category_distribution", categoryDistribution);
